requirements.txt
torch
matplotlib
transformers
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
huggingface_hub
faiss-cpu
faiss-gpu
sentence-transformers
#accelerate
bitsandbytes
optimum
pyyaml


#networkx # general purpose
#PyCG # static call graph
#graphviz # general?
#pycallgraph2 # for exact runtime call